Move source-target features into external-source modules
